sort an array
