+++ 912_select_sort.py
@@ -1,9 +1,11 @@
+# n = len(arr)
 # time: O(n**2)
 # space: O(1)
 from typing import List
 
 
-class Solution(object):
+class Solution:
+    
     def sortArray(self, arr: List[int]) -> List[int]:
         for ind, ele in enumerate(arr):
             min_ind = min(range(ind, len(arr)), key=lambda x: arr[x])
